Removed concurrency from ping package by moving PingAllNodes along with its unit tests to worker package and unexporting it
