+++ internal/ping/ping_test.go
@@ -92,51 +92,6 @@ func TestPingNodeTimeout(t *testing.T) {
 	cancelTimeout()
 }
 
-// TestPingAllNodes makes sure that PingAllNodes returns the errors we expect based on valid/invalid input
-// We also make sure that the number of successes and failures reported matches what we expect
-func TestPingAllNodes(t *testing.T) {
-	var successCount, failureCount int
-	numGood := 2
-	numBad := 1
-	ctx := context.Background()
-	if testing.Verbose() {
-		t.Logf("Ping all nodes - %d good, %d bad", numGood, numBad)
-	}
-	pingChannel := PingAllNodes(ctx, goodNode(""), badNode(badhost), goodNode(""))
-	for n := range pingChannel {
-		if n.Err != nil {
-			failureCount++
-		} else {
-			successCount++
-		}
-	}
-	if successCount != numGood || failureCount != numBad {
-		t.Errorf("Expected %d good, %d bad nodes.  Got %d good, %d bad instead.", numGood, numBad, successCount, failureCount)
-	}
-}
-
-// TestPingALlNodesTimeout pings a series of nodes with a 1 ns timeout and makes sure we get the timeout error we expect
-func TestPingAllNodesTimeout(t *testing.T) {
-	// Timeout
-	if testing.Verbose() {
-		t.Log("Running timeout test")
-	}
-	timeoutCtx, cancelTimeout := context.WithTimeout(ctx, time.Duration(1*time.Nanosecond))
-	pingChannel := PingAllNodes(timeoutCtx, badNode(""), badNode(badhost))
-	for n := range pingChannel {
-		if n.Err != nil {
-			lowerErr := strings.ToLower(n.Err.Error())
-			expectedMsg := "context deadline exceeded"
-			if lowerErr != expectedMsg {
-				t.Errorf("Expected error message to be %s.  Got %s instead", expectedMsg, lowerErr)
-			}
-		} else {
-			t.Error("Expected some timeout error.  Didn't get any")
-		}
-	}
-	cancelTimeout()
-}
-
 func TestParseAndExecutePingTemplate(t *testing.T) {
 	node := "mynode"
 	expected := []string{"-W", "5", "-c", "1", node}
